Base Pixels + Movement Flag Refactor (#2293)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
https://github.com/tgstation/tgstation/pull/54652
https://github.com/tgstation/tgstation/pull/54706 (partial)
https://github.com/tgstation/tgstation/pull/55444 (and fixes)

This is going to break things? But its also going to fix things

Fixes floating effects.

Base pixels allow for things with starting offsets to be animated and
pixel shifted more accurately.

closes #2312

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->