Hammer time (#2914)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Made it so hammers use defines for their force and gave them appropiate
integrity variables.
Forgefiend is now a hammer, so you can smith and craft with it. You can
also smith with a wooden mallet too. Ideally I would make it so smithing
with a wooden mallet puts you at a disadvantage in terms of quality, but
I would rather do it later, perhaps in a seperate PR.

## Why It's Good For The Game

Making the hammers use defines makes things more consistent, which is
better I believe.
Malum is the god of craftsmanship, so making his weapon able to craft is
I think, both a cool idea and fits in with the setting.
Made the wooden mallet able to smith so you can smith from "scratch".
Using defines is I think good for the codebase, as it makes things
cleaner and clearer.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

:cl:
add: Made the Forgefiend a hammer, you can craft and smith with it now
add: Wooden mallet can now be used to smith with
code: hammers use defines for their force and integrity, set some maces
integrity to use their defines
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

Co-authored-by: PotatoTomahto <76505086+PotatoTomahto@users.noreply.github.com>