Soap changes (#2955)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

There's a minor change to the description of the blessed mood, but
that's unrelated. "I feel a soothing." doesn't make sense in english I
think.
The clean and clean plus mood effects were increased to 20 and 30
minutes respectively
Their stress reduction however is both at -1, this is because clean and
clean_plus are both applied in the case of using the herbal soap, for a
total of -2 stress reduction.
Cleaning yourself in normal soap will give you the normal clean mood
effect.
Made it so soap is consumed when dilluted in a bucket of water.

## Why It's Good For The Game

Firstly, the mood length was buffed to encourage people to clean
themselves
The normal soap was also made to give out the clean effect as to, again
encourage people to clean themselves, but also to allow horcs to clean
themselves, which is what prevents their stink to radiate out of them.
I found it a bit silly that soap wasn't consumed at all when dilluted in
a bucket of water

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

:cl:
balance: Cleaning people with normal soap applies the clean mood effect,
which will remove the stink of horcs. Cleaning yourself with herbal soap
applies a greater effect, and being cleaned by someone with herbal soap
will apply the clean buff
balance: Soaps now lose a bit of use when dilluted in a bucket of water,
it is still way better than using it to clean floors directly
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: PotatoTomahto <76505086+PotatoTomahto@users.noreply.github.com>