Blessed needle can repair clothes (#2910)
## About The Pull Request
Changes the Desc of the Needle to fix typos and makes it able to repair
clothes, it's a needle after all.

## Changelog
:cl:
balance: Made the Blessed Needle capable of repairing clothes.
/:cl:

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->