Fixes #2986 and a bit more (#3015)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Basically the title
Fixes the #2986 issue, but also does some minor changes to the essence
harvester and the essence jars.
Basically, instead of plopping the node onto the ground when removing a
node from the harvester, instead you need to remove it via the
containment jar.
Removed the install_node and remove_node procs as they were no longer
being used after my ~~terrible~~ changes.

## Why It's Good For The Game

Making things work is nice. Also I think my changes make sense. You
extract the node with the funky jar, you put the node into the machine
with the funky jar, and you need to extract the node from the harvester
with the funky jar.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
fix: Fixed essence harvester menu. You also need to use the jar to now
remove nodes from it.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->