Minor Artificier (and blacksmith) recipe changes (#3045)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Lead bullets now cost tin to make, there are four made, and the lockpick
recipe is properly named
Also made lead bullets craftable by blacksmiths, because I think it
should

## Why It's Good For The Game

Tin is used to make tin/lead platters, forks and spoons. So why not lead
bullets.
Crossbow bolts and Bow arrows were already craftable by blacksmiths, why
not lead bullets, which are basically simple spheres.
Contrast with a Crossbow which would obviously require an artificier's
engineering, and wouldn't (and isn't) craftable by a blacksmith.
Also the lockpick recipe makes 3 lockpicks, not 2. Hence the rename.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: Lead balls are now crafted with tin at the artificier table
add: Lead balls can be crafted with tin by blacksmiths
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->