Encumbrance changes. Plate armor nerfs pt 1 (#2905)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Plate armor has been king of the hill for too long, hopefully this and
my next plate armor nerf PR will allow them to be less one-man murder
machines and more bulwarks that thrive with the support of lighter
troops, and die when they don't have escorts with them.

- Makes encumbrance affect how fast you get up from resting on the
ground
- Gives plate armor a further multiplier on this (fullplate armor is
hard to bend in)
- Removes the strength check on all encumbrance effects (you already
check endurance to calculate the denominator in the )
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
People in plate armor shouldn't be sprinting around like they own the
place, and encumbrance doesn't really exist or do anything right now due
to the strength check.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog
:cl:
balance: Makes fullplate and encumbered people get up slower
balance: Removes strength check from encumbrance and makes carryweight
scale on either STR or CON depending on whats higher.
balance: 50% encumbrance now disables sprinting from 70%
/:cl:
<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->



<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->