Nerfs Dendor Lesser Miracle (#3073)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Title

## Why It's Good For The Game

Besides the Dendor miracle, the others are situational or depends on
stuff like day or night, traits etc, while the Dendor Lesser miracle
will always give the max buff essentially doubling the damage if they
are in a Forest or area with nature, this makes so a Briar, Templar or
Acolyte of Dendor are extremely powerful against undead, while clearly
the forest buff should continue, being strong as it is right now makes
no sense.

This will reduce the likes of liches, skeletons, vampires,vampire lords
getting soloed by a briar in the midle of the forest spamming lesser
miracle on them. (Even tho i agree it can be an skill issue, it still
the strongest lesser miracle with the easiest condition to fulfill, 99%
of the times getting the max possible buff from the nature)

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
bal: Nerfs Dendor Lesser Miracle
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->