Weapon defines and stuff (#2930)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Mostly code clean up and that's it.
Consider this a complementary to my other pr,
https://github.com/Monkestation/Vanderlin/pull/2914
At that time, I didn't know its even worse
Also renamed MEDIOCHRE_PARRY to MEDIOCRE_PARRY because MEDIOCHRE Is not
a fucking word and no one bothered to change it
Also also, courtphysician cane is now a walking stick, because it's a
cane, aka a walking stick

## Why It's Good For The Game

Weapons should use defines because there are there for a reason. I hate
this codebase sometimes.
Also the golden/silver scabbards sell for more now, because why wouldn't
they, they are made from gold and silver
Courtphyscian's cane is a walking cane, so it should be able to support
people, hence it now counts as a walking stick

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

:cl:
add: court physician's cane is now a walking stick
code: made weapons use defines as they should
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: grungussuss <96586172+grungussuss@users.noreply.github.com>