fixing oiling up debuff and making it less unbearable (#3091)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Oiling up shouldn't be terrible

## Why It's Good For The Game

The slip chance for being oiled up was too high. Oiling up is supposed
to be good for wrestling, but I fail to see how tripping every 5 steps
with a hard stun makes it good for wrestling in any way whatsoever. So I
halved the percentage to slip. This makes oiling up more viable, but
still punishing when you do actually slip. I hardly see this becoming
meta.
## Changelog
qol: halved slipping percentage

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->