Lakkarian Weapons (#2887)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This pull request aims to gradually introduce the nation of Lakkari to
the world of Psydonia, a nation inspired by Nubia, Egypt, and the
Vatican City. Im doing this because I believe African culture,
mythology, and weaponry are underrepresented in medieval fantasy. I
decided weaponry would be a good way for me to start this introduction.

## Why It's Good For The Game

I believe this will add more weapon variety to the game, and could be
used for future roles relating to Lakkari.
## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

:cl:
add: Added new weapons with smithing recipes.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: NPC1314 <NPC1314@protonmail.com>
Co-authored-by: InsightfulParasite <109536843+InsightfulParasite@users.noreply.github.com>
Co-authored-by: Dorodomki <128561204+Dorodomki@users.noreply.github.com>
Co-authored-by: NPC1314 <110836368+NPC1314@users.noreply.github.com>
Co-authored-by: Optimism333 <81007356+Optimism333@users.noreply.github.com>
Co-authored-by: 1N7Slayer <145714829+1N7Slayer@users.noreply.github.com>
Co-authored-by: BadAtThisGame <79063506+BadAtThisGame302@users.noreply.github.com>
Co-authored-by: Jeff Watchson <johncenatoutourourou@gmail.com>
Co-authored-by: lavillastrangiato <63124803+lavillastrangiato@users.noreply.github.com>
Co-authored-by: La Villa Strangiato <mangodeity@gmail.com~>
Co-authored-by: DimmaDunk <Rottmaster14@gmail.com>
Co-authored-by: anything86 <43508099+anything86@users.noreply.github.com>
Co-authored-by: SELFHELL <99771731+SELFHELL@users.noreply.github.com>
Co-authored-by: cornercube <81207946+cornercube@users.noreply.github.com>
Co-authored-by: Cannibal Hunter <135169022+CannibalHunter@users.noreply.github.com>
Co-authored-by: ThePooba <81843097+ThePooba@users.noreply.github.com>
Co-authored-by: dwasint <82520990+dwasint@users.noreply.github.com>
Co-authored-by: bowlofcereal <50346600+bowlofcereal@users.noreply.github.com>
Co-authored-by: Pirouette <57296132+listerla@users.noreply.github.com>
Co-authored-by: TheMrBakit <maximebeaulegobeil@hotmail.com>
Co-authored-by: Cadet00 <65573420+Cadet00@users.noreply.github.com>
Co-authored-by: Aegledis <aegledis@gmail.com>
Co-authored-by: herma <49307226+hermaplusplus@users.noreply.github.com>
Co-authored-by: Very Angry Monochrome Man <112135997+Catalaria@users.noreply.github.com>
Co-authored-by: LesPirouettes <pirouette@tutamail.com>
Co-authored-by: Aria <125321780+Aegledis@users.noreply.github.com>
Co-authored-by: DAKKA-WAAAGH <34065421+DAKKA-WAAAGH@users.noreply.github.com>
Co-authored-by: Dova <thatoneguywithshotgun@gmail.com>
Co-authored-by: Dorodomki <wingusbingus577@gmail.com>
Co-authored-by: BadAtThisGame <jonhcenatoutourourou@gmail.com>
Co-authored-by: cornercube <specyalic3@hotmail.com>
Co-authored-by: WWWWW-Git <Yeloumafre@gmail.com>
Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>