A follow up PR to #3060 (#3076)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

~~DO NOT MERGE UNTIL #3060 IS MERGED~~
The PR changes various numbers to work in tandem with the fact that
ounces are gone, making most reagent numbers divisible by 2 and/or 5,
for ex. the waterskin has now a volume of 60 units instead of 64
I tried making it so the balance of the game isn't disturbed, but
obviously that couldn't be always avoided
The changes are as follows:
Buckets and Pots hold 100 units instead of 99, combined with the fact
that;
Soups and stews now require 25 units instead of 33, you can now make 4
different soups/stews at once in a pot
Cups hold 25 units instead of 24
Waterskin holds 60 units, and transfers units in amounts of 5 or 10
alchemical vials hold 30 units and transfer 10 units per swig
normal vials hold 40 units and transfer 5 units per swig
pure moondust now holds 15 units of the pure moondust reagent instead of
18
herbal remedies and salves now require 25 and 20 water/ethanol,
respectively
bottles can transfer in amounts of 5 or 10 (did you know they already
had volume set as 70? A number that's not divisible by 3? I hate
roguetown and zeth so much its unreal)
teapots can hold 100 units now
Carafes hold 100 units and transfer 10 units

## Why It's Good For The Game

the old values were made with ounces in mind, which were basically units
divided by 3
these ones, I believe, make more sense

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: Tweaked the volume of a lot of reagent containers, shouldn't be
noticeable changes
add: However, soups and stews now require 25 units of water instead of
33, so you can make 4 "batches" of stew in one full pot
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->