Silver Changes (#3084)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

- Give Silver enchantment for most of the silver items ingame.

- Changed abit the ziliqua description to fit it not being like other
silver stuff

- Change the checked for the psycross in your neck, it now also checks
for the psycross in your wrist.

- If you have a psycross on your neck or wrist, it can protect you from
a werewolf infection on a 50% rate, without a psycross it will be a 100%
as always.

- Vampires can't suck the blood of people that got silver on them(also
siring), to prevent vampires transforming and having psycrosses on their
wrist, silver armor or the likes.

## Why It's Good For The Game

We had many items like silver armor, noc helmet, silver masks that
didn't had enchantment and took silver to make, just didn't made sense
to not have it, same as other things as cups and etc.

Makes the Psycross slightly more useful since arm protection is worth
trading more than neck protection, werewolf chance to fail an infect on
someone with the psycross also make it more useful besides being biten
once and it is over.

Vampires should not be able to wear silver at all but the code didn't
cover if someone with silver got converted, anyways if you get to the
point of drinking the blood of someone they are most likely dying or
gave up, this will just prevent vampires running around with psycrosses
and carrying silver weapons by magically not being their bane anymore.
(Once transformed it would proc a single time the worn thing but it
can't kill and i believe one shotting someone over being transformed and
having silver on them is not a good mechanic either)

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: Silver Enchants to the missed silver stuff
bal: Vampires can't suck the blood if someone has something of silver on
them
bal: Werewolves have a 50% chance of failing to infect someone if that
person has a psycross on their wrist or neck.
bal: Psycross on your wrist protects from things only the psycross on
your neck would.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->