Minor changes to artificier recipes (#3107)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Mostly recipe name changes/tweaks to clarify things
also made water arrows/bolts craftable, it was in a todo comment that
was never done, classic
regarding vial arrows/bolts, I haven't figured out a way to make them
transfer the reagents from the vial to the arrow/bolt, so that's that
also I made the crossbow recipe require 2 fibres instead of 1, balance
slop
also added a comment to the artibackpack, mostly for those who will be
reading the code in the future and have a split-second confusion (like I
did) if the artibackpack even does anything

## Why It's Good For The Game

Clarity when it comes to names. That's a good thing right?
Regarding the crossbow. To me it felt very odd that the crossbow only
requires 1 fibre, given how thin the damned things are. Ropes felt too
thick, but since one rope is made from 3 fibres, I decided to make it so
the crossbow requires 2.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: water arrows/bolts are now craftable by artificers
qol: Clarified the recipe names of artificier craftable items
balance: Crossbows now require TWO fibres instead of ONE. Vanderlin
balance in shambles
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->